Style handlers replaced namespace with modules.
