+++ lib/graph3d/Graph3d.js
@@ -8,7 +8,7 @@ var Camera       = require('./Camera');
 var Filter       = require('./Filter');
 var Slider       = require('./Slider');
 var StepNumber   = require('./StepNumber');
-var SH           = require('./StyleHandler');    // Not called 'StyleHandler' to avoid clash with namespace
+var StyleHandler = require('./StyleHandler');
 
 
 /**
@@ -172,30 +172,33 @@ Graph3d.prototype._convert3Dto2D = function(point3d) {
  *                   camera
  */
 Graph3d.prototype._convertPointToTranslation = function(point3d) {
-  var ax = point3d.x * this.scale.x,
-    ay = point3d.y * this.scale.y,
-    az = point3d.z * this.scale.z,
+  var cameraLocation = this.camera.getCameraLocation();
+  var cameraRotation = this.camera.getCameraRotation();
 
-    cx = this.camera.getCameraLocation().x,
-    cy = this.camera.getCameraLocation().y,
-    cz = this.camera.getCameraLocation().z,
+  var ax = point3d.x * this.scale.x;
+  var ay = point3d.y * this.scale.y;
+  var az = point3d.z * this.scale.z;
+
+  var cx = cameraLocation.x;
+  var cy = cameraLocation.y;
+  var cz = cameraLocation.z;
 
   // calculate angles
-    sinTx = Math.sin(this.camera.getCameraRotation().x),
-    cosTx = Math.cos(this.camera.getCameraRotation().x),
-    sinTy = Math.sin(this.camera.getCameraRotation().y),
-    cosTy = Math.cos(this.camera.getCameraRotation().y),
-    sinTz = Math.sin(this.camera.getCameraRotation().z),
-    cosTz = Math.cos(this.camera.getCameraRotation().z),
+  var sinTx = Math.sin(cameraRotation.x);
+  var cosTx = Math.cos(cameraRotation.x);
+  var sinTy = Math.sin(cameraRotation.y);
+  var cosTy = Math.cos(cameraRotation.y);
+  var sinTz = Math.sin(cameraRotation.z);
+  var cosTz = Math.cos(cameraRotation.z);
 
   // calculate translation
-    factor0 = (sinTz * (ay - cy) + cosTz * (ax - cx)),
-    factor1 = (cosTy * (az - cz) + sinTy * factor0),
-    factor2 = (cosTz * (ay - cy) - sinTz * (ax - cx)),
+  var factor0 = (sinTz * (ay - cy) + cosTz * (ax - cx));
+  var factor1 = (cosTy * (az - cz) + sinTy * factor0);
+  var factor2 = (cosTz * (ay - cy) - sinTz * (ax - cx));
 
-    dx      = cosTy * factor0 - sinTy * (az - cz),
-    dy      = sinTx * factor1 + cosTx * factor2,
-    dz      = cosTx * factor1 - sinTx * factor2;
+  var dx      = cosTy * factor0 - sinTy * (az - cz);
+  var dy      = sinTx * factor1 + cosTx * factor2;
+  var dz      = cosTx * factor1 - sinTx * factor2;
 
   return new Point3d(dx, dy, dz);
 };
@@ -249,12 +252,12 @@ Graph3d.prototype._setBackgroundColor = function(backgroundColor) {
     strokeWidth = 0;
   }
   else if (typeof(backgroundColor) === 'object') {
-    if (backgroundColor.fill !== undefined)    fill = backgroundColor.fill;
-    if (backgroundColor.stroke !== undefined)    stroke = backgroundColor.stroke;
+    if (backgroundColor.fill        !== undefined) fill        = backgroundColor.fill;
+    if (backgroundColor.stroke      !== undefined) stroke      = backgroundColor.stroke;
     if (backgroundColor.strokeWidth !== undefined) strokeWidth = backgroundColor.strokeWidth;
   }
   else if  (backgroundColor === undefined) {
-    // use use defaults
+    // use defaults
   }
   else {
     throw 'Unsupported type of backgroundColor';
@@ -685,8 +688,6 @@ Graph3d.prototype.setData = function (data) {
  * @param {Object} options
  */
 Graph3d.prototype.setOptions = function (options) {
-  var cameraPosition = undefined;
-
   this.animationStop();
 
   if (options !== undefined) {
@@ -736,7 +737,9 @@ Graph3d.prototype.setOptions = function (options) {
       'zStep',
       'zMax',
       'valueMin',
-      'valueMax'
+      'valueMax',
+      'axisColor',
+      'gridColor'
     ];
 
     // retrieve simple parameter values
@@ -757,20 +760,18 @@ Graph3d.prototype.setOptions = function (options) {
 
     if (options.showLegend      !== undefined) this.defaultShowLegend = options.showLegend; 
     if (options.tooltip         !== undefined) this.showTooltip       = options.tooltip;
-    if (options.cameraPosition  !== undefined) cameraPosition         = options.cameraPosition;
     if (options.backgroundColor !== undefined) this._setBackgroundColor(options.backgroundColor);
 
-    if (cameraPosition !== undefined) {
+    if (options.cameraPosition !== undefined) {
+      var cameraPosition = options.cameraPosition;
       this.camera.setArmRotation(cameraPosition.horizontal, cameraPosition.vertical);
       this.camera.setArmLength(cameraPosition.distance);
     }
 
     // colors
-    if (options.axisColor !== undefined)        this.axisColor = options.axisColor;
-    if (options.gridColor !== undefined)        this.gridColor = options.gridColor;
     if (options.dataColor) {
       if (typeof options.dataColor === 'string') {
-        this.dataColor.fill = options.dataColor;
+        this.dataColor.fill   = options.dataColor;
         this.dataColor.stroke = options.dataColor;
       }
       else {
